+++ packages/core/svelte/use-completion.ts
@@ -2,8 +2,9 @@ import { useSWR } from 'sswr'
 import { Readable, get, writable } from 'svelte/store'
 
 import { Writable } from 'svelte/store'
+
+import type { UseCompletionOptions, RequestOptions } from '../shared/types'
 import { createChunkDecoder } from '../shared/utils'
-import { UseCompletionOptions } from '../shared/types'
 
 export type UseCompletionHelpers = {
   /** The current completion result */
@@ -13,7 +14,10 @@ export type UseCompletionHelpers = {
   /**
    * Send a new prompt to the API endpoint and update the completion state.
    */
-  complete: (prompt: string) => Promise<string | null | undefined>
+  complete: (
+    prompt: string,
+    options?: RequestOptions
+  ) => Promise<string | null | undefined>
   /**
    * Abort the current API request but keep the generated tokens.
    */
@@ -47,6 +51,7 @@ export function useCompletion({
   id,
   initialCompletion = '',
   initialInput = '',
+  credentials,
   headers,
   body,
   onResponse,
@@ -76,7 +81,7 @@ export function useCompletion({
   const isLoading = writable(false)
 
   let abortController: AbortController | null = null
-  async function triggerRequest(prompt: string) {
+  async function triggerRequest(prompt: string, options?: RequestOptions) {
     try {
       isLoading.set(true)
       abortController = new AbortController()
@@ -88,10 +93,15 @@ export function useCompletion({
         method: 'POST',
         body: JSON.stringify({
           prompt,
-          ...body
+          ...body,
+          ...options?.body
         }),
-        headers: headers || {},
-        signal: abortController.signal
+        headers: {
+          ...headers,
+          ...options?.headers
+        },
+        signal: abortController.signal,
+        credentials
       }).catch(err => {
         throw err
       })
@@ -157,8 +167,11 @@ export function useCompletion({
     }
   }
 
-  const complete = async (prompt: string) => {
-    return triggerRequest(prompt)
+  const complete: UseCompletionHelpers['complete'] = async (
+    prompt: string,
+    options?: RequestOptions
+  ) => {
+    return triggerRequest(prompt, options)
   }
 
   const stop = () => {